requirements.txt
Flask==3.0.3
gunicorn==20.1.0
python-dotenv==1.0.1
mysql-connector-python==9.1.0
pytz==2024.1
requests==2.31.0
tenacity==8.2.2
pandas==2.2.0
pyarrow==15.0.0

# Google Cloud SDKs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.91.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.14.0
google-cloud-core>=2.4.1

# Google Sheets
gspread==6.2.0

# HTML parser
beautifulsoup4==4.12.3
